html/semantics/scripting-1/the-script-element/scriptexecute-events/028.html

<!DOCTYPE html>
<html>
  <title>afterscriptexecute for DOM inserted script</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    events = [];
    var t = async_test();
    document.addEventListener("afterscriptexecute",
      function x(e) {
        events.push(e.target);
      }, false);
    var script = document.createElement("script");
    script.textContent = "events.push('script ran');";
    document.body.appendChild(script);
 </script>
 <script>
   t.step(function(){assert_array_equals(events, ["script ran", document.getElementsByTagName("script")[3],
                                                  document.getElementsByTagName("script")[2],])});
   t.done();
 </script>
